Refactor: update package declarations after module split
- Update package names in common module
- Update package names in model module
- Update package names in repository module
- Update domain test imports
