+++ core/domain/src/commonTest/kotlin/pt/dourobats/app/core/domain/model/LanguageTest.kt
@@ -1,4 +1,4 @@
-package pt.dourobats.app.core.domain.model
+package pt.dourobats.app.core.model
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
