+++ core/model/src/commonMain/kotlin/pt/dourobats/app/core/model/SessionStatus.kt
@@ -1,4 +1,4 @@
-package pt.dourobats.app.core.domain.model
+package pt.dourobats.app.core.model
 
 /**
  * Represents the session state.
