+++ core/model/src/commonMain/kotlin/pt/dourobats/app/core/model/Sport.kt
@@ -1,4 +1,4 @@
-package pt.dourobats.app.core.domain.model
+package pt.dourobats.app.core.model
 
 /**
  * Represents a sport that can be associated with training sessions.
